+++ R/srs.R
@@ -16,19 +16,14 @@
 #' used absolute distances when squared distances are more commonly used in the
 #' world of distance and association metrics.
 #'
-#' There are potentially instances in which the conceivable range of
-#' ratings/attachments (i.e. your two vectors) are not observed. In the case of
-#' applications to alliance data, this is almost an impossibility. Every state,
-#' by assumption, is maximally committed to defending itself. There will
-#' assuredly be cases in which there is no commitment to another state in the
-#' data (either for reasons of disinterest or enmity, though the first calls
-#' into question what a 0 should communicate and the latter betrays the
-#' interesting complexity of alliances). Thus, the minimum and maximum, one
-#' assumes, will always be observed in the alliance data. Perhaps the same could
-#' be said for UN voting data, though I couldn't rule out the possibility that
-#' there is a dyad out there for which both states never voted "yes" or "no".
-#' That would have implications for the range in the denominator of the formula.
-#' You can override that by hard-setting the range in the `range` argument.
+#' There will sometimes be instances, assuredly with alliances, where not all
+#' categories are observed. For example, the toy example I provide of Germany
+#' and Russia in 1914 includes no 2s. In the language of "ratings", the "rating"
+#' of 2 was available for Germany and Russia in 1914 but neither side used it.
+#' The `levels` argument allows you to specify the full sequence of values that
+#' could be observed, even if none were. It probably makes the most sense to
+#' always use this argument, even if the default behavior operates as if you
+#' won't.
 #'
 #' The function subsets to complete cases of the two vectors for which you want
 #' an S score. If weights are included, the function further subsets to complete
@@ -42,8 +37,14 @@
 #'
 #' If it were my call to make, I'd caution against the IR standard of using
 #' the composite index of national capabilities  (CINC) as a weight on the
-#' calculation of the S statistic. Conceptually, weighting by capabilities tries
-#' to capture some kind of "importance" quantity. Related to the familiar
+#' calculation of the S statistic. I expand a bit on this line of thinking on my
+#' blog here:
+#'
+#' \url{https://svmiller.com/blog/2026/06/alliances-weighting-foreign-policy-similarity/}
+#'
+#' The following was my original entry into this documentation file and I'll
+#' keep it as is for posterity. Conceptually, weighting by capabilities tries to
+#' capture some kind of "importance" quantity. Related to the familiar
 #' application of alliances, this would prioritize those states that could
 #' conceivably bring more to the battlefield. In practice, this adds one
 #' anachronism to another. Capabilities, as measured, are basically a nineteenth
@@ -84,9 +85,6 @@
 #' doesn't assume the weights sum to 1. You'll see how in the denominator of
 #' the formula.
 #'
-#' Weights are only applicable to absolute distances. If you specify a weight
-#' variable with `distances = 'squared'`, the function will ignore your weights.
-#'
 #' In applications to the Correlates of War system, as far as I am aware, there
 #' are no CoW states for which there isn't a CINC estimate. If, for some reason,
 #' a CINC score (or some other weight) is missing, the cases are dropped
@@ -95,6 +93,9 @@
 #' If weights are supplied, the weights must match the length of either `x1` or
 #' `x2`. The function builds in an implicit assumption that the weights are a
 #' column in the data frame you're using.
+#'
+#' The function will proportionalize your weights to sum to 1 if they do not sum
+#' to 1 already.
 #'
 #'
 #'
@@ -105,16 +106,15 @@
 #' see note in details section.
 #' @param weights a vector of weights. Defaults to NULL for creating unweighted
 #' S statistics
-#' @param range defaults to NULL, but an optional vector that forces the range
-#' to be a certain value. If NULL, the function calculates a range based on the
-#' maximum and minimum values observed across both `x1` and `x2`. See details
-#' section for more.
+#' @param levels defaults to NULL, but an optional vector that defines the full
+#' sequence of values that could be observed in `x1` and `x2`. If NULL, the
+#' function looks for observed values.
 #'
 #' @examples
 #'
-#' srs(gmyrus14$gmy, gmyrus14$rus, distances = 'absolute')
-#' srs(gmyrus14$gmy, gmyrus14$rus, distances = 'squared')
-#' srs(gmyrus14$gmy, gmyrus14$rus, distances = 'absolute', weights = gmyrus14$syscap)
+#' srs(gmyrus14$gmy, gmyrus14$rus, distances = 'absolute', levels = c(0:3))
+#' srs(gmyrus14$gmy, gmyrus14$rus, distances = 'squared', levels = c(0:3))
+#' srs(gmyrus14$gmy, gmyrus14$rus, distances = 'absolute', weights = gmyrus14$syscap, levels = c(0:3))
 #'
 #' @references
 #'
@@ -125,7 +125,7 @@
 #' @importFrom stats complete.cases
 #' @export
 
-srs <- function(x1, x2, distances = 'absolute', weights = NULL, range = NULL) {
+srs <- function(x1, x2, distances = "absolute", weights = NULL, levels = NULL) {
 
   if(length(x1) != length(x2)) {
     stop("`x1` and `x2` are not the same length.")
@@ -135,12 +135,21 @@ srs <- function(x1, x2, distances = 'absolute', weights = NULL, range = NULL) {
     stop("`weights` must be the same length as `x1` and `x2` if you're going to provide it.")
   }
 
-  # force complete cases, just in case
-  if(is.null(weights)) {
-  completetf <- complete.cases(x1, x2)
+  if (is.null(levels)) {
 
-  x1 <- x1[completetf]
-  x2 <- x2[completetf]
+    use.these.levels <- sort(unique(c(x1, x2)))
+
+  } else {
+
+    use.these.levels <- levels
+
+  }
+
+  if(is.null(weights)) {
+    completetf <- complete.cases(x1, x2)
+
+    x1 <- x1[completetf]
+    x2 <- x2[completetf]
 
   } else {
 
@@ -152,41 +161,95 @@ srs <- function(x1, x2, distances = 'absolute', weights = NULL, range = NULL) {
 
   }
 
-  if(is.null(range)) {
+  ## 1. Absolute Distances ----
+  if(distances == "absolute") {
 
-    levs <- sort(unique(c(x1, x2)))
-    diff <- max(levs) - min(levs)
-
-  } else {
-
-    diff <- range
-
-  }
-
-  if(distances == 'squared') {
-
-    num <-  sum((x1 - x2)^2)
-    denom <- length(x1)*(diff^2)
-
-  } else if(distances == 'absolute') {
 
     if(is.null(weights)) {
+      ## * Absolute Distances, No Weights -----
 
-      num <- sum(abs(x1 - x2))
-      denom <- length(x1)*diff
+      # Calculated observed S...
+      tab <- table(factor(x1, levels = use.these.levels),
+                   factor(x2, levels = use.these.levels))
 
-    } else {
+      o <- prop.table(tab)
 
-      num <- sum(abs(x1 - x2)*weights)
-      denom <- sum(weights)*diff
+      d <- abs(row(o) - col(o))
+      dd <- max(use.these.levels) - min(use.these.levels) # nrow(o) - 1 # nrow(o) - 1
 
+      S <- 1 - 2*sum(o*d)/dd
+
+
+    } else { # Okay, so you have weights... cue Shania Twain...
+
+      ## * Absolute Distances, with Weights ----
+
+      # Calculate observed S, with weights...
+      o <- xtabs(weights ~
+                   factor(x1, levels = use.these.levels) +
+                   factor(x2, levels = use.these.levels))
+
+      # Check if o sums to 1... it must...
+
+      if(sum(o) != 1) {
+        o <- o/sum(o)
+      }
+
+      rmarg <- rowSums(o)
+      cmarg <- colSums(o)
+
+      d  <- abs(row(o) - col(o))
+      dd <- max(use.these.levels) - min(use.these.levels) # nrow(o) - 1
+
+      S <- 1 - 2*sum(o*d)/dd
 
     }
 
   }
 
-  s <- 1 - 2*(num/denom)
+  ## 2. Squared Distances ----
+  if(distances == "squared") {
 
-  return(s)
+    if(is.null(weights)) {
+      ## * Squared Distances, No Weights -----
+
+      # Calculated observed S...
+      tab <- table(factor(x1, levels = use.these.levels),
+                   factor(x2, levels = use.these.levels))
+
+      o <- prop.table(tab)
+
+      d <- abs(row(o) - col(o))^2
+      dd <- (max(use.these.levels) - min(use.these.levels))^2 # nrow(o) - 1
+
+      S <- 1 - 2*sum(o*d)/dd
+
+    } else { # Okay, so you have weights... cue Shania Twain...
+      ## * Squared Distances, With Weights -----
+
+      # Calculate observed S, with weights...
+      o <- xtabs(weights ~
+                   factor(x1, levels = use.these.levels) +
+                   factor(x2, levels = use.these.levels))
+
+      # Check if o sums to 1... it must...
+
+      if(sum(o) != 1) {
+        o <- o/sum(o)
+      }
+
+      rmarg <- rowSums(o)
+      cmarg <- colSums(o)
+
+      d  <- abs(row(o) - col(o))^2
+      dd <- (max(use.these.levels) - min(use.these.levels))^2 # nrow(o) - 1
+
+      S <- 1 - 2*sum(o*d)/dd
+
+    }
+
+  }
+
+  return(S)
 
 }